requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
marshmallow==3.20.2
Flask-Marshmallow==1.0.0
geopy==2.4.1
requests==2.31.0
fastapi==0.115.2
uvicorn==0.29.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.11
googlemaps==4.10.0
geopy==2.4.1
polyline==2.0.2
stripe==7.11.0
supabase==1.2.0